+++ mescal/typevars.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from mescal.kpis.kpi_base import KPI
     from mescal.kpis.aggs import OperationOfTwoValues
     from mescal.kpis.kpi_collection import KPICollection
-    from mescal.visualizations.folium_viz_system.base_system import ResolvedStyle, StyleResolver
+    from mescal.visualizations.folium_viz_system.base_viz_system import ResolvedStyle, StyleResolver
 
 FlagType = TypeVar('FlagType', bound='FlagTypeProtocol')
 DatasetType = TypeVar('DatasetType', bound='Dataset')
