+++ mescal/visualizations/folium_viz_system/viz_line_text_overlay.py
@@ -6,8 +6,12 @@ from folium.plugins import PolyLineTextPath
 from shapely import LineString
 
 from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem, KPIDataItem
-from mescal.visualizations.folium_viz_system.base_system import ResolvedStyle, StyleResolver, StyleMapper, \
-    FoliumObjectGenerator
+from mescal.visualizations.folium_viz_system.base_viz_system import (
+    ResolvedStyle,
+    StyleResolver,
+    StyleMapper,
+    FoliumObjectGenerator,
+)
 
 
 @dataclass
