+++ mescal/visualizations/folium_viz_system/viz_text_overlay.py
@@ -3,11 +3,15 @@ from typing import Callable, Type
 
 import folium
 
-from mescal.visualizations.folium_viz_system.simple_generators import TooltipGenerator, PopupGenerator
+from mescal.visualizations.folium_viz_system.element_generators import TooltipGenerator, PopupGenerator
 from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem, KPIDataItem
 
-from mescal.visualizations.folium_viz_system.base_system import ResolvedStyle, StyleResolver, StyleMapper, \
-    FoliumObjectGenerator
+from mescal.visualizations.folium_viz_system.base_viz_system import (
+    ResolvedStyle,
+    StyleResolver,
+    StyleMapper,
+    FoliumObjectGenerator,
+)
 
 
 @dataclass
